requirements.txt
python 3.12.1
flask-3.1.2
flask_session-0.8.0